pyproject.toml
[tool.poetry]
name = "cw-6"
version = "0.1.0"
description = ""
authors = ["venediktova-kate <venediktova.kat@gmail.com>"]
readme = "README.md"
packages = [{include = "cw_6"}]

[tool.poetry.dependencies]
python = "^3.10"
django = "^4.2.2"
psycopg2-binary = "^2.9.6"
djangorestframework = "^3.14.0"
drf-spectacular = "^0.26.2"
django-cors-headers = "^4.0.0"
djoser = "^2.2.0"
djangorestframework-simplejwt = "^5.2.2"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
